# remove bank conflict in iter grouped block reduction