--- a/runtime/block_reduction.cu
+++ b/runtime/block_reduction.cu
@@ -48,6 +48,10 @@ __device__ void blockReduce(
       index_utils::maskedOffset<!X_REDUCE, !Y_REDUCE, !Z_REDUCE>(
           threadIdx, blockDim);
 
+  // number of reductions per block
+  unsigned int reduction_num =
+      index_utils::maskedSize<!X_REDUCE, !Y_REDUCE, !Z_REDUCE>(blockDim);
+
   // smem_offset is the offset into shared memory for the current thread.
   // To ensure coalesced access to shared memory, we need to ensure
   // each transaction is accessing a contiguous block of 128 bytes.
@@ -64,6 +68,39 @@ __device__ void blockReduce(
   unsigned int smem_offset = threadIdx.x + threadIdx.y * blockDim.x +
       threadIdx.z * blockDim.x * blockDim.y;
 
+  // The peer stride is the distance between the current element and its
+  // reduction peer, it depends on the reduction dimension.
+  constexpr int num_redu_dims = (int)X_REDUCE + (int)Y_REDUCE + (int)Z_REDUCE;
+  // reduction in 3 dimensions, XYZ, stride is 1
+  unsigned int peer_stride = 1;
+  if (num_redu_dims == 1) {
+    // Reduction only in 1 dimension, X or Y or Z
+    // e.g. inner or outer reduction
+    // If X_REDUCE, reducing in neighbor cols in smem, peer_stride is 1
+    // If Y_REDUCE, reducing in neighbor rows in smem, peer_stride is blockDim.x
+    // If Z_REDUCE, reducing in neighbor planes in smem, peer_stride is
+    // blockDim.x * blockDim.y
+    peer_stride = X_REDUCE ? 1
+        : Y_REDUCE         ? blockDim.x
+                           : blockDim.x * blockDim.y;
+  } else if (num_redu_dims == 2) {
+    // Reduction in 2 dimensions, only one dimension is not reduced, !X, !Y, !Z
+    // If !Z_REDUCE, merge XY, reducing neighbor cols, peer_stride is 1
+    // If !X_REDUCE, merge ZY, reducing neighbor rows, peer_stride is blockDim.x
+    // If !Y_REDUCE, if blockDim.y == 1, merge XZ, peer_stride is 1, otherwise
+    //               different strides for different reduction stages, change
+    //               smem layout based on reduction_idx and reduction_tid, may
+    //               have stride access and bank conflict, rare case, may happen
+    //               for batch norm doing multiple reductions per block.
+    if (!Y_REDUCE) {
+      smem_offset = blockDim.y > 1
+          ? reduction_idx * reduction_size + reduction_tid
+          : smem_offset;
+    } else {
+      peer_stride = !Z_REDUCE ? 1 : blockDim.x;
+    }
+  }
+
   // Initialize shared memory
   if (read_pred) {
     shared_mem[smem_offset] = inp_val;
@@ -76,14 +113,17 @@ __device__ void blockReduce(
   int np2 = 1 << (31 - __clz(reduction_size));
 
   if (reduction_tid < np2 && reduction_tid + np2 < reduction_size) {
-    reduction_op(shared_mem[smem_offset], shared_mem[smem_offset + np2]);
+    reduction_op(
+        shared_mem[smem_offset], shared_mem[smem_offset + np2 * peer_stride]);
   }
   block_sync::sync<Aligned>();
 
   // loop peel the final iteration to save one syncthread for the end
   for (int factor = np2 / 2; factor > 1; factor >>= 1) {
     if (reduction_tid < factor) {
-      reduction_op(shared_mem[smem_offset], shared_mem[smem_offset + factor]);
+      reduction_op(
+          shared_mem[smem_offset],
+          shared_mem[smem_offset + factor * peer_stride]);
     }
     block_sync::sync<Aligned>();
   }
@@ -92,7 +132,7 @@ __device__ void blockReduce(
     T result = out;
     reduction_op(result, shared_mem[smem_offset]);
     if (reduction_size > 1) {
-      reduction_op(result, shared_mem[smem_offset + 1]);
+      reduction_op(result, shared_mem[smem_offset + peer_stride]);
     }
     out = result;
   }
@@ -154,139 +194,146 @@ __device__ void blockIterGroupedReduce(
   static_assert(
       N == 2 || N == 4 || N == 8 || N == 16,
       "N should be a valid vectorization factor, one of (2, 4, 8, 16)!");
+  static_assert(
+      !X_REDUCE && Y_REDUCE && !Z_REDUCE, "Only support reduction in Y dim!");
 
-  bool should_write =
-      index_utils::maskedIsZero<X_REDUCE, Y_REDUCE, Z_REDUCE>(threadIdx);
+  bool should_write = threadIdx.y == 0;
+  unsigned int reduction_size = blockDim.y;
+  unsigned int reduction_tid = threadIdx.y;
 
-  unsigned int reduction_size =
-      index_utils::maskedSize<X_REDUCE, Y_REDUCE, Z_REDUCE>(blockDim);
+  // In shared memory, each row has 128 bytes, if sizeof(T) * N = 32 bytes, each
+  // row has 128 / 32 = 4 threads. Each transaction can only load data from one
+  // row, with a max of 16 bytes per thread. So the total bytes per transaction
+  // is 4 x 16 = 64 bytes which is only half of the maximum 128 bytes per
+  // transaction. we should change the layout from [TIDy, TIDx, N] to [N/4,
+  // TIDy, TIDx, 4]
+  constexpr unsigned int total_loads =
+      sizeof(T) * N / 16 > 1 ? sizeof(T) * N / 16 : 1;
+  constexpr unsigned int elements_per_load =
+      16 / sizeof(T) > N ? N : 16 / sizeof(T);
 
-  unsigned int reduction_tid =
-      index_utils::maskedOffset<X_REDUCE, Y_REDUCE, Z_REDUCE>(
-          threadIdx, blockDim);
+  // assume TIDy is the reduction dimension, TIDx is the iteration dimension
+  // TIDz is not used
+  unsigned int peer_stride = elements_per_load * blockDim.x;
 
-  unsigned int reduction_idx =
-      index_utils::maskedOffset<!X_REDUCE, !Y_REDUCE, !Z_REDUCE>(
-          threadIdx, blockDim);
+  unsigned int smem_offset_inter = blockDim.x * blockDim.y * elements_per_load;
+  unsigned int smem_offset_intra =
+      (threadIdx.y * blockDim.x + threadIdx.x) * elements_per_load;
 
-  // In shared memory, each row has 128 bytes, if sizeof(T) * N = 32 bytes, each row has 128 / 32 = 4 threads.
-  // Each transaction can only load data from one row, with a max of 16 bytes per thread.
-  // So the total bytes per transaction is 4 x 16 = 64 bytes which is only half of the
-  // maximum 128 bytes per transaction. we should change the layout from [TIDy, TIDx, N] to 
-  // [N/4, TIDy, TIDx, 4]
-  constexpr unsigned int total_loads = sizeof(T) * N / 16 > 1 ? sizeof(T) * N / 16 : 1;
-  constexpr unsigned int elements_per_load = 16 / sizeof(T) > N ? N : 16 / sizeof(T);
+// load to [total_loads] sections of shared memory
+#pragma unroll
+  for (unsigned int i = 0; i < total_loads; ++i) {
+    loadGeneric<T, elements_per_load>(
+        shared_mem + smem_offset_inter * i + smem_offset_intra,
+        const_cast<T*>(inp_val) + i * elements_per_load);
+  }
+  block_sync::sync<Aligned>();
 
-  if(true){
-    unsigned int smem_offset_inter = blockDim.x * blockDim.y * blockDim.z * elements_per_load;
-    unsigned int smem_offset_intra = (threadIdx.z * blockDim.x  * blockDim.y + threadIdx.y * blockDim.x + threadIdx.x) * elements_per_load;
-    
-    // load to [total_loads] sections of shared memory
-    #pragma unroll
+  // Reduce down to nearest power of 2 for the tree reduction:
+  // Perform parallel reduction for each element in the array
+  int np2 = 1 << (31 - __clz(reduction_size));
+  if (reduction_tid < np2 && reduction_tid + np2 < reduction_size) {
+    // vectorized load from smem to regs
+    T self[N];
+    T peer[N];
+#pragma unroll
     for (unsigned int i = 0; i < total_loads; ++i) {
+      int self_offset = smem_offset_inter * i + smem_offset_intra;
+      int peer_offset = self_offset + np2 * peer_stride;
       loadGeneric<T, elements_per_load>(
-          shared_mem + smem_offset_inter * i + smem_offset_intra,
-          const_cast<T*>(inp_val) + i * elements_per_load);
+          self + i * elements_per_load, shared_mem + self_offset);
+      loadGeneric<T, elements_per_load>(
+          peer + i * elements_per_load, shared_mem + peer_offset);
     }
-    block_sync::sync<Aligned>();
+// reduction
+#pragma unroll
+    for (int i = 0; i < N; ++i) {
+      reduction_op(self[i], peer[i]);
+    }
+// write self back to smem
+#pragma unroll
+    for (unsigned int i = 0; i < total_loads; ++i) {
+      int self_offset = smem_offset_inter * i + smem_offset_intra;
+      loadGeneric<T, elements_per_load>(
+          shared_mem + self_offset, self + i * elements_per_load);
+    }
+  }
+  block_sync::sync<Aligned>();
 
-    // Reduce down to nearest power of 2 for the tree reduction:
-    // Perform parallel reduction for each element in the array
-    int np2 = 1 << (31 - __clz(reduction_size));
-    if (reduction_tid < np2 && reduction_tid + np2 < reduction_size) {
+  // Tree reduction
+  for (int factor = np2 / 2; factor > 1; factor >>= 1) {
+    if (reduction_tid < factor) {
       // vectorized load from smem to regs
       T self[N];
       T peer[N];
-      #pragma unroll
+#pragma unroll
       for (unsigned int i = 0; i < total_loads; ++i) {
         int self_offset = smem_offset_inter * i + smem_offset_intra;
-        int peer_offset = self_offset + np2 * elements_per_load * blockDim.x;
-        loadGeneric<T, elements_per_load>(self + i * elements_per_load, shared_mem + self_offset);
-        loadGeneric<T, elements_per_load>(peer + i * elements_per_load, shared_mem + peer_offset);          
+        int peer_offset = self_offset + factor * peer_stride;
+        loadGeneric<T, elements_per_load>(
+            self + i * elements_per_load, shared_mem + self_offset);
+        loadGeneric<T, elements_per_load>(
+            peer + i * elements_per_load, shared_mem + peer_offset);
       }
-      // reduction
-      #pragma unroll
+// reduction
+#pragma unroll
       for (int i = 0; i < N; ++i) {
         reduction_op(self[i], peer[i]);
       }
-      // write self back to smem
-      #pragma unroll
+// write self back to smem
+#pragma unroll
       for (unsigned int i = 0; i < total_loads; ++i) {
         int self_offset = smem_offset_inter * i + smem_offset_intra;
-        loadGeneric<T, elements_per_load>(shared_mem + self_offset, self + i * elements_per_load);
-      }
-    }
-    block_sync::sync<Aligned>();
-
-    // Tree reduction
-    for (int factor = np2 / 2; factor > 1; factor >>= 1) {
-      if (reduction_tid < factor) {
-        // vectorized load from smem to regs
-        T self[N];
-        T peer[N];
-        #pragma unroll
-        for (unsigned int i = 0; i < total_loads; ++i) {
-          int self_offset = smem_offset_inter * i + smem_offset_intra;
-          int peer_offset = self_offset + factor * elements_per_load * blockDim.x;
-          loadGeneric<T, elements_per_load>(self + i * elements_per_load, shared_mem + self_offset);
-          loadGeneric<T, elements_per_load>(peer + i * elements_per_load, shared_mem + peer_offset);             
-        }
-        // reduction
-        #pragma unroll
-        for (int i = 0; i < N; ++i) {
-          reduction_op(self[i], peer[i]);
-        }
-        // write self back to smem
-        #pragma unroll
-        for (unsigned int i = 0; i < total_loads; ++i) {
-          int self_offset = smem_offset_inter * i + smem_offset_intra;
-          loadGeneric<T, elements_per_load>(shared_mem + self_offset, self + i * elements_per_load);
-        }
-      }
-      block_sync::sync<Aligned>();
-    }
-
-    // last reduction
-    if (should_write && write_pred) {
-      // init result
-      T result[N];
-      #pragma unroll
-      for (int i = 0; i < N; ++i) {
-        result[i] = out[i];
-      }
-  
-      // copy first element to result
-      T self[N];
-      #pragma unroll
-      for (unsigned int i = 0; i < total_loads; ++i) {
-        int self_offset = smem_offset_inter * i + smem_offset_intra;
-        loadGeneric<T, elements_per_load>( self + i * elements_per_load, shared_mem + self_offset);
-      }
-      #pragma unroll
-      for (int i = 0; i < N; ++i) {
-        reduction_op(result[i], self[i]);
-      }
-  
-      // reduction of the 2nd last element
-      if(reduction_size > 1){
-        T peer[N];
-        #pragma unroll
-        for (unsigned int i = 0; i < total_loads; ++i) {
-          int peer_offset = smem_offset_inter * i + smem_offset_intra + elements_per_load * blockDim.x;
-          loadGeneric<T, elements_per_load>( peer + i * elements_per_load,  shared_mem + peer_offset);
-        }
-        #pragma unroll
-        for (int i = 0; i < N; ++i) {
-          reduction_op(result[i], peer[i]);
-        }
-      }
-      #pragma unroll
-      for (int i = 0; i < N; ++i) {
-        out[i] = result[i];
+        loadGeneric<T, elements_per_load>(
+            shared_mem + self_offset, self + i * elements_per_load);
       }
     }
     block_sync::sync<Aligned>();
   }
+
+  // last reduction
+  if (should_write && write_pred) {
+    // init result
+    T result[N];
+#pragma unroll
+    for (int i = 0; i < N; ++i) {
+      result[i] = out[i];
+    }
+
+    // copy first element to result
+    T self[N];
+#pragma unroll
+    for (unsigned int i = 0; i < total_loads; ++i) {
+      int self_offset = smem_offset_inter * i + smem_offset_intra;
+      loadGeneric<T, elements_per_load>(
+          self + i * elements_per_load, shared_mem + self_offset);
+    }
+#pragma unroll
+    for (int i = 0; i < N; ++i) {
+      reduction_op(result[i], self[i]);
+    }
+
+    // reduction of the 2nd last element
+    if (reduction_size > 1) {
+      T peer[N];
+#pragma unroll
+      for (unsigned int i = 0; i < total_loads; ++i) {
+        int peer_offset =
+            smem_offset_inter * i + smem_offset_intra + peer_stride;
+        loadGeneric<T, elements_per_load>(
+            peer + i * elements_per_load, shared_mem + peer_offset);
+      }
+#pragma unroll
+      for (int i = 0; i < N; ++i) {
+        reduction_op(result[i], peer[i]);
+      }
+    }
+#pragma unroll
+    for (int i = 0; i < N; ++i) {
+      out[i] = result[i];
+    }
+  }
+  block_sync::sync<Aligned>();
 }
 
 // Use the same pred for both reads and writes