… runtime/grid_reduction.cu