**Issues:** 
(1) similar to #2203, when vectorized by 8, each thread holds 32 bytes of data, the write to smem is not coalesced
(2) similar to #2204, calculating smem offset using `reduction_tid` and `reduction_idx` leads to stride access for outer reduction.
(3) r/w to smem during the reduction process is not vectorized.

**Fix:** 
(1) similar to #2203, when vectorized by 8, split data into two sections. Each thread in each section has 16 bytes.
(2) similar to #2204, calculating smem offset using index of threads within a block.
(3) ensure vectorized r/w to smem during the reduction process.

**Other minor changes:**
renamed function name from `blockIterGroupedReduce` to `blockIterGroupedYdimReduce` since it is only used in outer reduction where Y dim is used for reduction. Also added check to codegen.

**Investigation of other reduction method:** 
also tried using warp reduction, the kernel latency is not changed much becuase the global memory access is the dominating factor instead of the reduction runtime function. See the [doc](https://docs.google.com/document/d/1IxbiylbeVz6FQ50viOTSn2-YKd5VBSdjmjYWbsTKYTg/edit#heading=h.16fz6k676m4k) for details.

**Results:**
(1) gelu bwd, speedup of this branch over main branch, larger is better. regression is due to the reduce of occupancy from 62% to 47% due to the increased register usage. 
![image](https://github.com/NVIDIA/Fuser/assets/116412316/7f9eb3ba-5763-46fe-a0fd-f8e979f8029e)
(2) outer reduction, similarly, the regression is due to reduced occupancy comes from increased register usage.
![image](https://github.com/NVIDIA/Fuser/assets/116412316/4d21e482-134e-4271-b458-e009878e2768)


**Why register usage slightly increased in some cases:**
(1) Ncu analysis shows at case 2048 x 20480, max live register increased from 46 to 49, the register assignd to each thread increased from 48 to 54. Seems like 48 is a critical value, if it needs more than 48 will bump up to the next value. 
(2) Even I remove optimization (1) and (3) and just keep optimization (2) `calculating smem offset using index of threads within a block`,  the register usage still increased from 48 to 54. Further check of PTX and SASS code showed the compiler is automatically vectorizing one of the shared memory load (compiler team says this is expected and is done by LSV, LoadStoreVectorizer).
(3) These small regressions will be fixed in the heuristic revision PR.

schedule paras at case 2048 x 20480:
```
Iteration Domain: blockIdx.x / threadIdx.x / multiple reductions per block / vectorize / factor 2
Inner Reduction Domain: cross block - threadIdx.y / cross grid - blockIdx.y / split grid dim / split grid dimension / unroll / factor 4
Launch Parameters: BlockDim.x = 160, BlockDim.y = 2, BlockDim.z = -1, GridDim.x = -1, GridDim.y = 16, GridDim.z = -1, Smem Size = 0
Compile Parameters: index_type = NotSet, maxrregcount = 255, enable_magic_zero = 1, enable_ptxas_verbose = 0
```