@liqiangxl 2024-05-21
> Naive question:
> 
> Given the description here `Ncu analysis shows at case 2048 x 20480, max live register increased from 46 to 49, the register assignd to each thread increased from 48 to 54.`
> 
> That doesn't look like a lot of extra register usage.
> 
> Though I'm not sure which schedule params would translate to the `N` used in the template.
> 
> ```
> Iteration Domain: blockIdx.x / threadIdx.x / multiple reductions per block / vectorize / factor 2
> Inner Reduction Domain: cross block - threadIdx.y / cross grid - blockIdx.y / split grid dim / split grid dimension / unroll / factor 4
> Launch Parameters: BlockDim.x = 160, BlockDim.y = 2, BlockDim.z = -1, GridDim.x = -1, GridDim.y = 16, GridDim.z = -1, Smem Size = 0
> Compile Parameters: index_type = NotSet, maxrregcount = 255, enable_magic_zero = 1, enable_ptxas_verbose = 0
> ```
> 
> Looking at the kernel, we are trying to maintain (mostly) two arrays of size N with type T. I thought that's going to be more register usage from the code change.

`N` corresponds to vectorization factor. The register change is small becuase the original code is unrolled. So, even without vectorization, if the compiler wont' reuse the register we still needs `N` registers to temperarally store the data loaded from shared memory. The attached figure is the PTX code corresponding to 
```
5322	#pragma unroll
5323	    for (int j = 0; j < elements_per_load; ++j) {
5324	        reduction_op(shared_mem[self_offset + j], shared_mem[peer_offset + j]);
5325	      }

```
in this case `elements_per_load = 2`. Two things are interesting:
(1) the load from shared memory `self_offset + j` is automatically vectorized (L504). This is a little surprise.
(2) then two loads from `peer_offset + j` and two adds, two stores to `self_offset + j`

![image](https://github.com/NVIDIA/Fuser/assets/116412316/a0b2598b-41d2-4922-bc97-eaf94ea10427)

For case with vectorized load, the cuda code is long but PTX code is short and clean. Just load self, load peer, sum and store.
![image](https://github.com/NVIDIA/Fuser/assets/116412316/3f6086cc-5fde-4539-9334-79584ffc3e83)