@liqiangxl 2024-05-22
> QQ regarding this part:
> 
> > (2) Even I remove optimization (1) and (3) and just keep optimization (2) calculating smem offset using index of threads within a block, the register usage still increased from 48 to 54.
> 
> I thought when we try to revert (1) and (3), we were seeing that perf regression was gone when benchmark against this PR?

That result is not correct. When I revert (1) and (3) I didn't make the code correct. After the correction, there is no performance change due to the auto vectorization added by LoadStoreVectorizer.