@liqiangxl 2024-05-28
The failed case in `nvfuser-ci/jit_nvfuser_bench_17_8/8 ` with msg `Error from segmentation group 2: c_ids.size() == 1 INTERNAL ASSERT FAILED at "/opt/pytorch/nvfuser/csrc/id_model/id_model.cpp":560,` seems to be caused by the accidently enalbed ID model. Can't reproduce locally after the latest main branch is merged into this branch. @naoyam