@liqiangxl 2024-06-04
Checked cpp benchmarks on H100, there are two regressions due to the increase of register usage:
```
NvFuserScheduler_BiasDropoutAddLayernormBwd1_fp32/512/128/1024/manual_time
NvFuserScheduler_BiasDropoutAddLayernormBwd1_tf32/512/128/864/manual_time
Fusion in these cases:
Inputs:
  T0_g[ iS0{i0}, iS1{i2}, iS2{i3} ], float
  T1_g[ iS27{i0}, iS28{i2}, iS29{i3} ], float
  T2_g[ iS30{i0}, iS31{i2}, bS8{1} ], float
  T3_g[ iS35{i0}, iS36{i2}, bS11{1} ], float
Outputs:
  T6_g[ rS18{i0}, rS19{i2}, iS20{i3} ], float
  T8_g[ rS24{i0}, rS25{i2}, iS26{i3} ], float
  T5_g[ iS37{i0}, iS38{i2}, iS39{i3} ], float

%kernel_math {
T6_g[ rS18{i0}, rS19{i2}, iS20{i3} ]
   = reduction( T0_g[ iS0{i0}, iS1{i2}, iS2{i3} ], op = add, initial value = float(0), allreduce = false )
T4_l[ iS32{i0}, iS33{i2}, iS34{i3} ]
   = T1_g[ iS27{i0}, iS28{i2}, iS29{i3} ]
   - T2_g[ iS30{i0}, iS31{i2}, bS8{1} ];
T5_g[ iS37{i0}, iS38{i2}, iS39{i3} ]
   = T4_l[ iS32{i0}, iS33{i2}, iS34{i3} ]
   * T3_g[ iS35{i0}, iS36{i2}, bS11{1} ];
T7_l[ iS21{i0}, iS22{i2}, iS23{i3} ]
   = T0_g[ iS0{i0}, iS1{i2}, iS2{i3} ]
   * T5_g[ iS37{i0}, iS38{i2}, iS39{i3} ];
T8_g[ rS24{i0}, rS25{i2}, iS26{i3} ]
   = reduction( T7_l[ iS21{i0}, iS22{i2}, iS23{i3} ], op = add, initial value = float(0), allreduce = false )
}
```
register usage:
```
With this PR: 70 registers per thread, occupancy = 25%
Before this PR: 64 registers per thread, occupancy = 50%
```
Performance:
Case | Main Branch (GB/s) | This PR-2205 (GB/s) | After Heuristic Revision, PR-2281(GB/s)
-- | -- | -- | --
512/128/1024 | 2570 | 2200 | 2670
512/128/864 | 2296 | 1974 | 2403

</div></b>