@liqiangxl 2024-06-04
Checked cpp benchmarks on A100, no regression detected.